Moved !say command to utils module
